Add a 'lines' combinator to the text package. Associated with PR #190 Suggested by @fthomas in association with proposed NIO based filesystem tools implemented in PR #190.
The lines(maxLineLength: Int): Process1[String, String] combinator will take a stream of Strings and repartition them by line endings. If `maxLineLength` is greater than zero and a line is determined to be longer than `maxLineLength` the Process will fail with an `java.lang.Exception`. The behavior with regard to new lines is indended to be identical to that of scala.io.Source.getLines.