Removed a performance optimization in Free#step.
The performance optimzation eagerly applied `Bind(Pure(x), f)`
while stepping. IIRC, this optimization gave ~20% performance
improvement in some benchmarks. Unfortunately, it is unprincipled. More
specifically, the optimization assumed that `f` is pure. However,
`Free.suspend` (and `Stream.suspend`) are implemented as
`pure(()).flatMap(_ => s)`, so we either need to introduce `suspend` as
a new primitive in `Free` (and `Stream` and maybe other places) or we
need to avoid the assumption that `f` is pure.

Note that the optimization allowed us to use `Either[Throwable,?]` in
the evaluation of pure streams in a stack safe way. Removal of the
optimization requires the effect type to provide its own stack safety.
For now, I've changed pure streams to use `Task` again, though I'm
considering introducing an `fs2.IO` type that provides an `Effect[IO]`
instance along with trampolined synchronous evaluation.