Fixed memory leak in FreeC
The cached `ViewL` structure in `FreeC` leaks memory (though I don't)
understand why exactly. Replacing the cached hard reference with a soft
reference helps, but heap usage grows until GC decides to clear soft
references. Replacing with a weak reference fixes that but it's not
clear how often we get a cache hit with a weak reference. Further, the
soft and weak reference versions have a performance impact on the common
case where stream evaluation does not benefit from ViewL caching --
about 5% in my tests.

To see the leak, run the `ConstantEvalSanityTest` with caching enabled.

In this PR, I've disabled caching completely. We can always add it back
if we determine why it leaks. Before adding it back, I'd like to have
some tests that demonstrate the cost of caching pays off in common use
cases.