Removed the Suspend node from the Stream algebra
We added the Suspend node back when we were trying to implement viewL
caching in FreeC. If viewL caches, then a Free.suspend(f) may be
evaluated and cached. However, suspend on Stream/Pull/Algebra relies on
always-evaluate semantics.

Now that we don't cache in FreeC, there's no reason for the Stream
algebra to have its own Suspend constructor.