Internal changes to the implementation of `translate`.
We make some changes to the internal Stream implementation, in the
Algebra file, to avoid using implicit classes.

- We remove the syntax extension implicit class `AlgScopeSyntax`,
  which was not being used anywhere.
- We remove the syntax extension `AlgEffectSyntax`, and extract its
  only method as an internal private function.
  We replace the only two uses needed for this function.

After removing the `AlgScopeSyntax` syntax extension, it turns out
that there is no place in the code in which the `AlgScope` sub-trait
of `Algebra` is used, so we can remove that one as well.