Changes and Syntactic Sugaring in the Internals
In the Algebra GADT, we make the type parameter `O` (which is the
return type of a `Stream` covariant. This allows us to define the
sub-trait `AlgEffect` as Algebra operations with `O=INothing`, in
other words, operations that are not concerned with output.
As a consequence, for those operations we can remove the middle
type parameter.

In the `Algebra.scala` file, we introduce several type aliases,
which we hope will help reading the code in here, and will make
more obvious the patterns between algebra operations.
- We use `ExprFA` as an alias for a `FreeC` of an `Algebra`.
- We use `ProgFA` as an alias for an `ExprFA` where the result is `Unit`.
  A `ProgFA` is precisely what is used to compute a Stream[F, A].
- We use `StepFA` to represent the output of a Step, which is:
  a chunk of output values, a middle result, and the continuation.

We also introduce a type aliases, for the whole `Internal` package,
to represent `Finalisers` and resource cleanups.