Changes and Syntactic Sugaring in the Internals
In the Algebra GADT, we make the type parameter `O` (which is the
return type of a `Stream` covariant. This allows us to define the
sub-trait `AlgEffect` as Algebra operations with `O=INothing`, in
other words, operations that are not concerned with output.
As a consequence, for those operations we can remove the middle
type parameter.