Remove the EitherSyntax class
We remove the EitherSyntax implicit class, used to provide a
right-biased flatMap for `Either`.
Although fs2 still supports the Scala 2.11, in truth the special
right-biased flatMap is used nowhere in `fs2`.
The only method being used was the `toOption`; and since that was
followed with `Option.get`, we can just use `.right.get` instead
by way of the Either.RightProjection class.