Remove the `O` parameter from CompileScope.
This commit removes the type parameter `O`, which is usually
associated to the output of a `Stream[F, O]`, from several internal
classes and types, such as `CompileScope`, its internal `State`,
the `InterruptContext, and the algebra operation `GetScope`.
We also remove it from the definitions of several auxiliary methods.

It turns out that `CompileScope` is not really concerned itself with
the emission of values of type `O`, but merely with the correct
handling of the shared resources in a common scope.
Removing the parameter `O` helps make this clearer.