Replace the use of ViewL.View with an "unvoldView"
The internals of the  `FreeC` had an ADT, which overlaps with `FreeC`,
called `ViewL`, used to represent `FreeC` expressions at which
evaluation in `FreeC` can stop. This evaluation proceeds by
unrolling left-nested `Bind` and unrolled, composing Bind-continuations
to the right, until it gets either a final result or  a `ViewL`.

It turns out, `ViewL` is not stored anywhere in any data structure,
and they are not used as type parameters anywhere.
Instead, a `ViewL` is emitted by using a method `viewL` (on the
`FreeC` class) and, in all usages, is immediately pattern-matched.

This suggests that we can remove the `ViewL` type altogether, and
replace the use of `viewL` by a higuer-order function `viewFold`,
which takes two continuations: one for results, and one for the
`ViewL.View` case (which represents a Bind-Eval).
In effect, we have unfolded the use of `ViewL`by two patterns.