Internal Algebra: unify translate0 and translateStep
The methods translate0 and translateStep were introduced in previous
versions, to deal with different variants of the translations.
- translate0 was introduced to remove translation from the Algebra,
- translateStep was introduced to support the new compiler-interpreter.

Looking at them, the only differences between these methods were:
- That translate0 was not parameterised on the output type X.
- That, for the algebra operation of output, the translate0 had a
  recursive call to translate0, whereas the translateStep method
  was performing a Cast to a result of the parametric type X.

Thus, we have unified these methods, by replacing the initial call to
`translate0` as a call to `translateStep` with `X = O`.

We also wrap the `translateStep` inside a `translateAux` function,
which is the main entry method of the recursive loop.
Since neither the `fK` nor the `Concurrent[G]` change in the
recursive calls of `translateStep`, we use a pattern useful in
recursive functions, of not passing in the recursive calls the
parameters, either type or value, that never change.