Make all Algebra subclasses private.
This PR modifies the Algebra object, to hide all the sub-classes of
that algebra. Thus, the implementation are not visible outside it,
and only in the algebra are those constructed/deconstructed.

To allow this, we cut from the implementation of `Pull.mapOutput` the
part that  generates a natural transformation, to a Algebra.mapOutput
operation.