Remove Type Parameters in FreeC and subclasses.
Using the fact that `FreeC[F[_], +R]` is covariant in O, we modify
its sub-classes `Fail` and `Interrupted`, so as to set `R` to INothing,
and then we can remove the type parameter from the class.
We apply the same change to some of the factory methods.

We also can remove the covery method.