Encapsulate CompileScope methods
We modify and narrow the API of the CompileScope class. We remove the
method "openAncestor", and hide as private "findSelfOrAncestor" and
"findSelfOrChild".
We expose two narrow methods, `descendsFrom` (a Boolean query), and
the `closeRelative` method, which includes the logic to find a related
scope, close it, and return the error.
We add documentation for these methods.

We rewrite the code of the `compileLoop-go` for the `CloseScope` node.
We use of the `closeRelative` method. This removes the `closeAndGo`.