CompileScope.open: rearrange operations
We reorganise the operations of the `open` method of the
`CompileScope` class.

We change `createCompileScope` from being an `Option` of the
`InterruptContext` to directly building the new scope.
This hides the `newScopeId` in this F-value.

This operation saves the allocations of a `Some` and a Tuple2 objects,
although now the `CompileScope` is now built regardless of path.
Not a big loss, since this is likely the frequent case.