Ugly performance Hack: Introduce No Exception.
There are many places in FS2 Internals where an `Option[Throwable]`,
or an `Either[Throwable, Unit]` is used to indicate presence or
absence of an error.

This has a memory cost: although the first one only has extra allocation
when there is an error, the second one always allocates either a
Right or a Left.

It turns out there is a zero-extra-allocations solution: since
`Throwable` is a trait and an open trait, if we want to represent new
cases we can just use an special sub-type, namely an object, that
extends from `Throwable`. This object will be an exception, but one
which actually represents the absence of Exceptions.

NoException is the good exception, the exception in the Exception
family, to indicate that, in that execution, all worked fine without
exception.