FreeC.Bind: use a method to represent a function.
In the FreeC datatype, the Bind constructor represents the case of a
bind or flatMap operation in a free monad constructor: it takes the
step FreeC[A], and the continuation of what to do with the result.
Currently, the implementation represents that continuation using a
value of type `Function1`.

However, in Scala there is another way to represent that a record has
a function attached: by declaring an abstract method, and using
inheritance as a form of parameterisation.

This commit changes the Bind case class to an abstract class with
an abstract `cont` method, which represents that continuation.
This avoids the allocation and indirection of `Function1` instances,