Optimisations to the `viewL` Loop.
This Commit introduces some optimisations to the View unrolling loop.
The View Unrolling loop takes an input `FreeC` and expands-evaluates
to remove all Binds on the left hand side, to end up with
either a Result, or a pair of an Eval and a continuation.

Before, when the unrolling loop found a case of a Bind with an inner
Bind on the left hand side, what it did was to
1) create a new Bind, made of the head of the inner one and the
   continuations of both the outer and the inner Bind; and
2) recurse on that Bind object, which immediately deconstructed it.

So, this would create a Bind object that it would throw away in the
next iteration, which is an uneeded allocation.

To remedy this inefficiency, we introduce an inner loop (@tailrec
function) that takes two parameters, the continuations and the heads
of the Binds. When a Bind-Bind case happens, it just recurs on the
left spine of the structure.

The loop may sometimes compose a continuation (a Bind-in-Bind),
and then apply that continuation to a intermediate result,
to obtain a FreeC that may be a Bind, which it then needs to unroll.
In this case, a Bind object is created to be immediately deconstructed.
To avoid this, we introduce a special sub-class of Function, called
>=> (a composition operator for monadic functions).
When the loop meets an intermediate Result, it checks to see if the
continuation is a  >=>, and if it is, it recurs in the inner loop.