Fix the Stack Overflows in creating a Bind-Bind.
The previous commit was causing some stack overflows in the inner cont
method, which was delegating to the `b.cont` method of other Bind.
Unlike functions, where the reference to the continuation was passed,
when using objects method calls are not "aliviated", and instead
create cost in the Stack.

To solve this problem, we add to the Bind class a new field, a
`delegate`, which points to another Bind object of same types.
This allows us to simulate a delegated call: when the main cont method
is called, it follows the chain of delegation (in a loop), without
adding extra Stack frames, until it reaches the non-delegated one.

Instead of unrolling the delegate fields when we call to the `cont`
method, we can directly unroll the first step upon creating the
object. In which case, every Bind created during the mk loop will
be referring to the first non-delegated bind.