ViewL: avoid using Function1 field, use method instead.
In line with the recent changes to Bind, we now change the View class,
(intended to represent an Eval + a continuation),
so as to turn the `next` function into a class method.

To handle the special case of an Eval, for which a "pureContinuation"
was introduced, we add a special subclass of View, PureView, which
uses the `._asFreeC` call. This subclass is kept private to the
View, so as to not raise the exhaustiveness check.