Make Chunk#scanLeft behaviour match standard collections
Changes Chunk#scanLeft to match the behaviour of the scanLeft
method on the standard collections, by including the zero element
and only including the final element once, while maintining the
existing behaviour of Chunk#scanLeftCarry.

Also adds tests for this behaviour, by comparing the result of
Chunk#scanLeft with List#scanLeft.