Force evaluation/access to the delegate
Before this commit, in the Bind-on-Bind case we were creating an outer
Bind (`nb`) whose `cont` method would create an inner `Bind`, and the
`cont` method of the inner `Bind` would call to b.delegate.
However, because the `delegate` is a value, this effectively meant
that a reference to the `b` object was kept in the closure of `cont`.
This was preventing the Garbage collector for reclaiming the object
`b`, and making it not possible to optimise that call, even though the
delegate is an immutable field.

In this commit, we force the evaluation of `b.delegate` expression,
and kept its value inside the new outer `Bind`, which allows us to
discard the `b`. This helps improve the performance in benchmarks.