Use back-to-front letters for intermediate Result types
The use of the `R` as the return type of the `mk` was a bit confusing
when dealing with the Bind-on-Bind case, in which we end up needing
two extra variables (for types of intermediate results of Binds).
Using letters Z, Y, and X, conveys the intuition of sequencing better.