Algebra: replace the R class with a CompileEnd trait
The `compileLoop` method has a system of nested executions,
which correspond to the Step algebra operations, in which
an inner stream is executed with a different CompileScope.
This was implemented with an inner ADT, using the word R
(for Result) and three cases:

- The compilation on a certain Scope is done, without result.
- The compilation is stopped with some output values, and the
  rest of the stream; or
- The compilation was interrupted.

This ADT is only used as a way of messaging from one part of the
method to another: objects created in one part are deconstructed in
another part.

This commit replaces the use of that ADT by a "object algebra",
a trait called "CompileEnd", whose methods correspond to the three
cases of that ADT.

The CompileEnd acts as a continuation in CPS: it indicates how the
execution of the stream should proceed when it reaches an end-point.

There are two (sub)-types of CompileEnd: the first one is used
for returning from the `compileLoop` method.
the second one is used when dealing with a Step instruction.