Move append operation to FreeC abstract class.
Given that append is probably very used Stream operator, it may make
sense to add a specialised operation to the FreeC class itseld,
as a variant of transformWith.

This may save the cost of creating a PartialFunction for any call.
It also removes a use of FreeC internals from `Stream`.