Unify FreeC and Algebra types.
This commit unifies the two core ADT of fs2: FreeC and Algebra.
FreeC is a free monad with catch and interruption.
Algebra represents a set of "abstract instructions" that may output
values of type O, or put a value of type R on a stack.
A Stream[F, O] consists of a `FreeC[Algebra[F, O, ?], Unit]`.

We have noted two interesting facts of this combination:
- In fs2, a FreeC was only used when its F[_] type was an Algebra
- The only place in which an Algebra object would appear in a FreeC
  was through the Eval class.

Based on these facts, and to simplify and reduce memory consumption
(as well as some complexity) of the Stream implementation, we
unify FreeC and Algebra, but turning "Algebra" into a subset of FreeC:

- We add to FreeC a third parameter, where the second parameter
  represents the output of a Stream.
- We turn the "Eval" class of FreeC into an abstract trait,
- We remove the Algebra trat, and turn all the former Algebra
  sub-classes into sub-classes of that Eval.

This change causes other changes:

- The `translate` function of the FreeC.Bind is no longer needed. This
  was needed to implement `mapOutput` as a natural transformation.
  Instead, we add the `mapOutput` function to the FreeC itself.
- We introduce another middle sub-class of FreeC, for the results.
  This is just a mere convenience, to reduce code elsewhere.
- Since classes and types are simplified, some factory methods or
  "smart constructors", needed to lift through types, are not needed.
  Such as Algebra.step, openScope, closeScope.