Stream: reimplement range without fold
The `Stream.range` operation, for creating a stream of integers from
an initial value to a final value in a step-by-step way, was
implemented using fold. Using a fold has a bit of extra laziness and
the creation of the `Some` and tuple objects, which is useful when
doing effectful operations, but not for outputting ints.

This commit makes a lighter implementation of Stream.range.