FreeC: turn the F parameter Covariant.
We change the definition of `FreeC`, and its several subclasses, to
be covariant on the kind-1 type `F[_]` parameter.

Since the F is covariant in FreeC, we can define the Result classes
as "Pure" cases of FreeC, that is to say those where the `F` parameter
is the `fs2.Pure` (an alias for Nothing).
We can also remove the `asFreeC` type-casting method, as well as a few
other "asInstanceOf" operations.

Now that FreeC is Covariant, in Stream and Pull we can unify the outer
type parameter `F` with the type parameter of the inner `FreeC`, so
that a Pull (or a Stream) are now just simple wrappers.
This allows to remove some internal uses of the `get` function.

Whenever we need to add some type information (over Pure or INothing),
we can just do that with plain Scala type annotation.

- Remove some uses of asInstanceOf

- Also, we can save some type annotations in the benchmarks.
  Remove some usages of the covary methods.

- Remove commented lines inside tut.
  Apparently Scala 2.13 console does not deal with them.

- TextSpec: issues involving pipes and variance

- Add Mima problems