Solves memory Leak
Reverts changes in commit 36b1279cda415ebe8bb3faa33cb9aa3e6c6697d9.

This commit changed the Compilation Loop was modified in a recent PR
to use a continuation and a stack of continuations. However, this
introduced a memory leak that could be seen running a simple:
`Stream.range(0, 2800934).fold(0L)(_ + _).compile.last`,
as described in Issue 1625.

This commit reverts those commits, and thus fixes that memory leak.