Dirty tricks and hacks to appease Scala 2.12. compiler.
When compiling on Scala 2.12, we were hitting on a compiler bug,
that manifests as a compiler crash, wherever we pattern-match
on the Step and OpenScope sub-classes of FreeC.

This bug is due to our FreeC now being a Generalised ADT, with
a type parameter that is covariant and of higuer-kind.
The specific characteristic why this fails for these two sub-classes
and not others is that, in these, the type-parameter F appears in the
fields of the case class.

While this bug has been fixed in Scala 2.13, in fs2 we still need to
adapt to work on Scala 2.12. To that end, we have to relax a bit
the type-safety of these classes, and introduce some `asInstanceOf`
to bridge the gaps left by that.