Stream/Pull: purge fromFreeC/get converters.
The internal methods `fromFreeC` and `get` of Pull and Stream were
needed at first to convert between the wrapped FreeC and them.
This conversion was needed due to the difference between FreeC
invariant type-parameter F[_], and Stream/Pull covariant +F[_].
Now that FreeC is also covariant on F[_], we can drop this converters.

Since these methods were private to fs2, but not entirely private,
it turns out that Mima detects them as missing methods.
We have to mark them as exceptions in the build