Unify FreeC and Pull: first attempt.
After removing `fromFreeC` and `get`, we have observed that many of
the operations in Pull were directly mapped to the inner FreeC,
whereas the others were based on those operations.
Also, given that `FreeC` and `Pull` have effectively the same
parameters, we can unify them.

this commit tries to unify `Pull` and `FreeC` under the hood,
by turning `Pull` into an abstract class (with abstract methods) and
`FreeC` into the only direct subclass of Pull.

The reason for using this way is to preserve some degree of
binary-compatibility, if there is some to be kept, and