A Stream can Pull anything: change Stream to `FreeC[F, O, IAny]`
A `Stream[F, O]` is defined as a wrap on a `FreeC[F, O, Unit]`.
The Unit type indicates, in types, that the "Result" type of the FreeC
computation that would yield the stream is of no consequence:
a Stream is a complete FreeC on its own, and its result is of no matter.

We can also express that idea, in types, by marking that return type
as an `Any`, the top type, which is another way to indicate that no
useful information can come off the Result of the inner FreeC.
One benefit is that we do not need spurious conversions on the FreeC
object just to convert it back-and-forth a Stream: Any thing will do.

To avoid problems with the Any poly-kindness, and much like with the
INothing alias, we use a special IAny alias of Any for the types.