FreeC: merge compile and compileLoop methods.
The `FreeC` compile logic was split in two top-level methods:
`compileLoop` deals with iterating the evaluation (compilation) until
a chunk of output is produced; whereas `compile` implements the outer
loop, which aggregates each chunk of outputs into the accumulated result.

This commit unifies these methods, by making both the outer and inner
loops into private methods of a `compile` method.

This reduces length, and has the small benefit of "inlining" the
generation of an F value at the end of the inner loop,
with the flatMap performed on it in the outer loop.