FreeC: remove polymorphic recursion in compile-go and Translate
The `Algebra` companion object contains two methods to interpret or
transform a `FreeC`: compile and translate. Both methods have an inner
recursive method, `go` and `translateStep` respectively, which takes
a generic type `X`, and which in the initial call (top level) always
coincides with the output type `O`.

The reason for this polymorphic recursion is to deal with the `Step`
class in the FreeC ADT: the Step contains an inner `FreeC`, the output
type of which cannot be tied to the `O` parameter of `FreeC`.

However, we can just assume that the output type of a Step's inner
stream is just the same as the `O`, and remove the polymorphic
parameter and polymorphic recursion in both methods,
by using a couple of `asInstanceOf` operations.