Weave translation into the language and the compilation.
Co-authored-by: Michael Pilquist <mpilquist@gmail.com>

The `translate[G[_]]` operation changes the effect type F[_] on which
a Stream is built or interpreted, by applying a natural transformation
to each F-effect in the  stream. Currently, translation is implemented
as a tree reconstruction, at the Pull free-monad algebra, so that a
new Pull object and its components are built during translation.
This was no longer satisfactory, given new changes in cats-effect.

This commit replaces the existing tree-rebuilding approach with one
that incorporates translation into the stream compilation.
We introduce a new Pull constructor, the `Translate` class, which
embeds a natural transformation and the stream to be translated.

=== Compile loop ===

We incorporate translation into the compile loop: as the stream is
unrolled, each action is interpreted and translated to the "final"
effect, the target of the compilation. To do this, we add in the
`go` method of the compile loop:

- A type parameter G[_], to indicate the effect the stream is "written
  in",  which is different from the "final" one.
- A parameter for the translation (~>) that we are folding the stream with.
- Each `Translate` instruction, we compose that extra translation.
- We handle stops (from the `Step` class) by translating the Stop
  directly and passing it to the top level.

The alternative would be to compose (through flatmaps and the like)
the actions of the inner algebra into a single G-action,
and only translate at the boundary of that inner stream.
This would not be a streaming solution, though, and would not work
with the intention of Step, which is to yield back control.

=== OpenScope ===

The OpenScope class no longer needs to provide an instance of
the `TranslateInterruptible` type class, to indicate if it was
interrupted  or not. Interruption is now completely handled by
the compilation, as part of "interpreting" or "running" the stream.

Before, `OpenScope` used an `Option` to indicate if interruption
was desired or not. Now, it is just a boolean.

=== Concurrent Instances ===