Pull - split compile-shift out of the Compile loop.
In the compile loop, the case for the `Step` class, includes some
logic to shifting from the current scope to another scope, by id.
Since the compileLoop is already large (and we may add more things
to it) it may be convenient to move some logic to the CompileScope
class.