Pull ViewL: we only apply it for streams.
The ViewL operation and trait, which represents the unrolling and
construction of a Pull object by unrolling until we reach a head,
is only ever used when interpreting or compiling a Stream.
Therefore, we can assume that it onlt works on Pull with Unit as
the result type, and remove that type Parameter.

To do this, we need to remove references to it from the Pull class
itself (which is parameterised on R). For that:

- We remove the class method `mapOutput` of the Pull class, and its
  implementation for several subclasses. Instead we write a new
  method with large match.

- We remove the `asHandler` method of the `Pull` class and directly
  expand it in the only place where it was used, at top of compilation.

- We remove the nested object ViewL, but make all its subclasses
  Private to the Pull object.