Pull - CompileLoop: reorganise into local methods.
Currently, the compileLoop is a large match statements, with one case
for each constructors of Pull, and in each case we have the full body.
This creates a small confusion of parameter names, and type names.
It also causes trouble due to the GADT handling in Scala.

Rather than doing that, we can create a few inner methods. Each method
can declare any type parameters of its own, or fix those that should
be determined by the type of constructor (as a good GADT would).