introduce Stream.paginate
When consuming resources from an API using pagination, it is usefull to signal at the same time that there are no more elements to read and to emit the last ones read.

For example, when iterating 10 by 10 elements, the last "page" can return 6 elements.
In the case, the `Stream` can emit those 6 elements and stop there, as we know that there are no more elements to read (`6 < 10`)