Pull - Replace OpenScope with InScope
The OpenScope pull constructor, to change the evaluation into a new
resource scope, was only used at one place, followed by a Pull flatMap
into the continuation. This continuation was intended to wrap an inner
stream, the one whose compilation is to run in the opened scope.

We replace the OpenScope class with an InScope class, which does
provide the inner stream. Most of the semantics is now moved to the
`goInScope` method, of evaluation. AS much as possible, we use
direct interpretation over Pull creation.