FlatMapOutput - Introduce Pull constructor.
We add a Pull subclass for the last major Stream operation, which is
to generate a stream from output of a stream (flatMap on Output).
This new `FlatMapOut` class is similar to the `MapOutput` class
from a recent PR, but the interpretation is different. Instead of
doing an interpretation in the `go` method of the compile Loop, as
another case, we have to handle this constructor in the View method.
The reason is that, because the `flatMapOutput` method before was
expanding into a Bind object, that broke the invariants that make
the `viewL` pull-unrolling method tail-recursive, and thus causing
several StackOverflows.

Instead, that expansion of the `Step` and the continuation is now
done in the `viewL` method, in which a `FlatMap` is directly turned
into a `View`, ending the loop unrolling.