Introduce constructor: FlatMapOutput
We introduce a new Pull constructor, for the Stream-flatMap (flatMap
on the output) operation. WE also introduce a smart constructor, for
a quick peeks.

The old body of output-flatMap is now moved as an auxiliary method of
the compile loop, called `innerFlatMapOutput`. We still keep using
Pull expansion before doing the interpretation, but this is now
bundled with the compile loop.

We also turn the `interruptBoundary` function into an auxiliary method
of the `innerFlatMapOutput`.