Pull.loop: change stream return type, from Option to Unit.
The `Pull.loop` had the return type `Pull[F, O, Option[R]]`, which
suggests that at the end of loop there would be some r or none.
But in truth, the loop ends when there is none, with none.
So we can just say that it ends with Unit.

This makes `Pull.loop` more similar to the _unfold_ method for building
lists from an `A => Option[B]` method. Here we can even build the
Stream from it.