Introduce `onCons` Pull transformation.
Looking through the code of `Stream`, we noticed that the `uncons`
method is used to implement a lot of the Stream combinators.
The `uncons` is often followed by that of `Pull.flatMap`.
Furthermore, the `uncons` implementation itself includes a `map`
operation, which adds several allocations to each operation

This commit introduces an `onCons` method, that combines the `uncons`
and the `flatMap`, to provide a better combinator.
The `onCons` function takes a function on what to do upon the first
"cons", which is to say, on the first Output from the stream.