Pull: split Step operation in two constructors.
The `Step` Pull constructor is used to turn a closed pull (a Stream)
into a pull that stops at the first `Output` node. It had an optional
scopeId parameter, to indicate if we shift to a scope or not.
The `None` case is used for the `uncons` operation, by mapping it.
The `Some` case is used for the `stepLeg`, mapping it to the StepLeg.

Since these are different usages with different data and different
end return type, it may be more Morganic to split them out.
We split the `Step` into two cases `StepCons` and `StepLeg`.

- StepCons is a step without shifting scopes, and it returns
  just the (option of) tuple of Chunk and tail pull.
- StepLeg is a step that is run in a separate scope, shifting there
  and back again. It gives back the `StepLeg` class.