Pull: use the carry-over notation.
Thinking about the right name for the 3rd pull parameter, the
one used to chain elements in `Pull.flatMap`.
We currently use the `R` letter for "return" or "result" or
"resource", but the first two are overloaded, and "resource"
has a rather different meaning.;

A better name may be "Carry": much like the carry-over operand of an
addition, which is passed to compute the rest of the operands,
a Pull leaves a carry-over result to be passed onto next results.

Using a separate letter (from `O`) allows us to use `D` for the
alternative carry-over parameter, rather than `R2`.