Question - Pull... eager append?
In `Pull` we have the `>>` method which performs a lazy append:
although the receiver `Pull` object is allocated before the call,
the `p2` or `post`  pull argument is passed lazily.

Lazy evaluation may have some allocation costs.
When the receiver / prefix pull is simple enough that, barring
interruption, the post is certain to be allocated.

The risk of this, of course, is that by preallocating the right-hand
side we may be leaking memory, for instance if the prefix is expanded
in an infinite pull. However, wouldn't that risk also exist with
thunks? Lazy evaluation may not be free either.