Pull: CloseScope and InScope - Refactorings.
The `boundToScope` function had  the effect of weaving an error,
from the inner pull, into the terminal result of the CloseScope.
However, that weaving can also be done in the `goCloseScope`
method, without the need to use the `transformWith` function.

The way of closing scope after  the `inScope` operation
only depends on how the inner streams ends, not the
other parameters of the `goInScope` method.

We simplify the text of the `sys.error`, since it is clear
by how the `CloseScope` is constructed that the case cannot happen.