Stream: Introduce foreach / foreachChunk functions.
The `evalMap` and `evalTap` operations can be used to perform
an effectful action on each element of the stream.
However, because Streams and lazy and we way want the consumer
of a stream to choose how long the effects of the stream are
performed, the evalMap (and evalTap) function need to emit
one singleton chunk of new output for each element on the input.

Once a stream is drained, though, the caller can no longer
control which of the effects are performed or not.
If evalMap is combined with drain, as we do at several places,
that is an unnecessary use of memory allocations.

To remedy that, this introduces the functions `foreach`
and `foreachChunk`, which take an effectful action and perform
it in every chunk or element of the input, perform them
eagerly, but emit nothing.