Pull - Compile MapOutput objects with a handler too.
The compilation of `MapOutput` is the only remaining `Pull` subclass
that is compiled by "rewriting" the `MapOutput`, which is to say
by reducing the `Pull` object to another `Pull` object, in the
`innerMapOutput` auxiliary method.

We introduce a new `MapOutR`, a subclass of Runner for the MapOutput
object. We remove the innterMapOutout rewrite method.

Add leak test for bug in https://github.com/typelevel/fs2/pull/2569