Pull - Remove BuildR Runners, use direct runners.
This "simplification", was tried before and rolled back because
it caused some Memory Leaks.
Inspecting a heapdump, we saw that the leaks were caused by the
compilation creating many objects of type `Bind`, `BindBind`, or
`DelegateBind`, in which the  `step` (the pull on the left of >>=)
was just a terminal.
To avoid this problem, we add a "smart constructor" method `bindView`,
used when creating a Pull by binding a Pull with a View.
This constructor, instead of always creating a new object, tries
to apply the evaluation-reduction instead.

One problem, however, is that this seems to cause one test on the
MonadError laws for Stream to fail, the stack-safety of `tailRecM`.

more